MD-4012 Make ChatContext serializable
ChatContext could not be serialized because public access to the fields
was not available.

I have started a unit test project and added a test that ensures that
the object can be serialized and deserialized in Json format. The test
is passing.

The Equals method has been added to support the comparison in the unit
tests and the GetHashCode method has been added because it is good
practice to add a GetHashCode when Equals has been overridden. These
were automatically generated by Visual Studio.